% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IKbandwidth.R
\name{IKbandwidth}
\alias{IKbandwidth}
\title{Imbens-Kalyanaraman Optimal Bandwidth Calculation}
\usage{
IKbandwidth(X, Y, cutpoint = NULL, verbose = FALSE, kernel = "triangular")
}
\arguments{
\item{X}{a numerical vector which is the running variable}

\item{Y}{a numerical vector which is the outcome variable}

\item{cutpoint}{the cutpoint}

\item{verbose}{logical flag indicating whether to print more information to the terminal. Default is \code{FALSE}.}

\item{kernel}{string indicating which kernel to use. Options are \code{"triangular"} (default and recommended), 
\code{"rectangular"}, \code{"epanechnikov"}, \code{"quartic"}, \code{"triweight"}, \code{"tricube"}, 
\code{"gaussian"}, and \code{"cosine"}.}
}
\value{
The optimal bandwidth
}
\description{
\code{IKbandwidth} calculates the Imbens-Kalyanaraman optimal bandwidth
for local linear regression in Regression discontinuity designs.
}
\author{
Drew Dimmery <\email{drewd@nyu.edu}>
}
\references{
Imbens, Guido and Karthik Kalyanaraman. (2009) "Optimal Bandwidth Choice for the regression discontinuity estimator," \emph{NBER Working Paper Series}. 14726. \url{http://www.nber.org/papers/w14726}
}

