% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd-package.R
\docType{package}
\name{rdd-package}
\alias{rdd}
\alias{rdd-package}
\title{Regression Discontinuity Estimation Package}
\description{
Regression discontinuity estimation package
}
\details{
\code{rdd} supports both sharp and fuzzy RDD
utilizing the \pkg{AER} package for 2SLS regression
under the fuzzy design. Local linear regressions are performed
to either side of the cutpoint using the Imbens-Kalyanamaran
optimal bandwidth calculation, \code{\link{IKbandwidth}}.
}
\author{
Drew Dimmery \email{drewd@nyu.edu}
}
\seealso{
\code{\link{RDestimate}}, \code{\link{DCdensity}}, \code{\link{IKbandwidth}},
\code{\link{summary.RD}}\code{\link{plot.RD}}, \code{\link{kernelwts}}
}

