\name{summary.RD}
\alias{summary.RD}
\title{Summarizing Regression Discontinuity Designs}
\usage{
  \method{summary}{RD} (object, ...)
}
\arguments{
  \item{object}{an object of class \code{"RD"}, usually a
  result of a call to \code{\link{RDestimate}}}

  \item{...}{unused}
}
\description{
  \code{summary} method for class \code{"RD"}
}
\author{
  Drew Dimmery <\email{drewd@nyu.edu}>
}

